test(LottoHandler): LottoHandler의 기능 테스트 후 수정
이전에 분리한 LottoHandler의 기능에 대해서
단위 테스트를 실행하고, 잘못된 로직을 수정함.

주요 수정 사항:
-
getResultOfLotto()에서 HashMap에 대한 초기화 문제 해결
- getRateOfResult()에서 Long
type을 Double type으로 수정
- LottoHandler의 상수 부분에서 잘못된 연산자(^)를 Math.pow()로
수정

+++ src/main/java/view/OutputView.java
@@ -24,7 +24,7 @@ public class OutputView {
                 .forEach(System.out::println);
     }
 
-    public static void printResultOfLotto(Map<LottoGrade, Integer> resultOfLotto, Long rateOfReturn) {
+    public static void printResultOfLotto(Map<LottoGrade, Integer> resultOfLotto, Double rateOfReturn) {
         System.out.println();
         System.out.println(RESULT_OF_LOTTO);
         System.out.printf(FIFTH + "%d" + "개\n", resultOfLotto.get(LottoGrade.FIFTH));
@@ -32,6 +32,6 @@ public class OutputView {
         System.out.printf(THIRD + "%d" + "개\n", resultOfLotto.get(LottoGrade.THIRD));
         System.out.printf(SECOND + "%d" + "개\n", resultOfLotto.get(LottoGrade.SECOND));
         System.out.printf(FIRST + "%d" + "개\n", resultOfLotto.get(LottoGrade.FIRST));
-        System.out.printf("총 수익률은 " + "%d" + "입니다.", rateOfReturn);
+        System.out.printf("총 수익률은 " + "%.2f" + "입니다.", rateOfReturn);
     }
 }
